Avoid showing values on cells that should be empty
For some reason, when returning None to a row's column value and mousing over
it, the treeview would get the latest non-None value read and display it on the
cell.

This fixes the problem by returning an empty string instead of None for
empty/invalid datetimes and also by returning a 1px transparent pixbuf instead
of None for images.