Cache images for datagrid
Because the way treeview and iconview works, the pixbufs that will
be displayed there are not cached. Because of that, everytime a
get_formatted_value is requested, a new transformation would be made.

That would happen everytime they wanted to redraw everything
(e.g. when scrolling the view or by simply hovering the mouse over a row)
or the first time a row was inserted (meaning n transformations to
expand a row with n children).

By caching the most recently used images, together with the visible
optimization range, we can do the transformation once and the cache
will be used for any next request.

This prevents the UI from freezing when expanding a row with too many
children, specially if those children are large images.