Allow to ensure the presence of a primary key
If ensure_primary_key is True and no primary key is found on the table,
datagrid will create one for it.

This is necessary for some things to work, like to select a single
row to call the callbacks when a row is selected/activated/etc,
to update the 'selected' column when it is present on the view
and other things.