Possible fix for a rare segmentation fault
Datagrid will segfault on some _very rare_ occasions. Tracking it with
gdb showed that the crash happened when calling IconView.choose_icon.

One of the possible reasons for this would be 2 threads accessing
the same object at the same time. So, the issue might be solved
by using another IconTheme instead of the default one.

Also, do a little optimization by caching the icon_filename based
on the file mime type.