Go back to using the default Gtk.IconTheme
Creating another IconTheme will make it sometimes fail to find
a given icon. It is better to use the default one.

To avoid having problems when multiple threads access that same
object, we will make sure that the main one will do that by using
the fallback image as the placeholder (the image that gets loaded
while the thread loads the real one).

This is actually nicer since the placeholder for an image will be
a image mimetype icon instead of a generic text one.